test: 이미지 테스트코드 분할
